feat: Optimize website structure and assets
- Cleaned up the file structure by removing unnecessary HTML and CSS files.
- Consolidated CSS files into a single `styles.css` to reduce HTTP requests.
- Consolidated JavaScript files into a single `app.js` to reduce HTTP requests.
- Optimized `index.html` for SEO and accessibility by updating the `<title>` and `<h1>` tags and adding a `<main>` tag.
