--- contacto.html
@@ -1,125 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Contacto - BellaDona</title>
-    <link rel="icon" type="image/x-icon" href="logo.ico">
-    <!-- Enlace a Google Fonts - Cookie -->
-    <link href="https://fonts.googleapis.com/css2?family=Cookie&display=swap" rel="stylesheet">
-    <!-- Enlace a Font Awesome para iconos -->
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
-    <!-- Enlace a nuestro archivo CSS -->
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-    <!-- Cabecera con navegación y selector de idioma -->
-    <header>
-        <div class="language-selector">
-            <!-- Selector de idiomas -->
-            <a href="#" class="active" data-lang="es">ES</a>
-            <a href="#" data-lang="ca">CA</a>
-            <a href="#" data-lang="en">EN</a>
-        </div>
-        <nav>
-            <a href="index.html" class="logo">BellaDona</a>
-            <ul class="menu">
-                <li><a href="index.html"><i class="fas fa-home"></i> <span class="lang-es">Inicio</span><span class="lang-ca">Inici</span><span class="lang-en">Home</span></a></li>
-                <li><a href="servicios.html"><i class="fas fa-cut"></i> <span class="lang-es">Servicios</span><span class="lang-ca">Serveis</span><span class="lang-en">Services</span></a></li>
-                <li><a href="reservas.html"><i class="far fa-calendar-alt"></i> <span class="lang-es">Reservas</span><span class="lang-ca">Reserves</span><span class="lang-en">Bookings</span></a></li>
-                <li><a href="sobre-mi.html"><i class="fas fa-user"></i> <span class="lang-es">Sobre mí</span><span class="lang-ca">Sobre mi</span><span class="lang-en">About me</span></a></li>
-                <li><a href="testimonios.html"><i class="fas fa-star"></i> <span class="lang-es">Testimonios</span><span class="lang-ca">Testimonis</span><span class="lang-en">Testimonials</span></a></li>
-                <li><a href="contacto.html" class="active"><i class="fas fa-envelope"></i> <span class="lang-es">Contacto</span><span class="lang-ca">Contacte</span><span class="lang-en">Contact</span></a></li>
-            </ul>
-            <div class="hamburger">
-                <i class="fas fa-bars"></i>
-            </div>
-        </nav>
-    </header>
-
-    <!-- Contenido principal -->
-    <main>
-        <div class="page-header">
-            <h1 class="page-title">
-                <span class="lang-es">Contacto</span>
-                <span class="lang-ca">Contacte</span>
-                <span class="lang-en">Contact</span>
-            </h1>
-            <p>
-                <span class="lang-es">¿Tienes alguna pregunta? Estamos aquí para ayudarte</span>
-                <span class="lang-ca">Tens alguna pregunta? Som aquí per ajudar-te</span>
-                <span class="lang-en">Do you have any questions? We are here to help you</span>
-            </p>
-        </div>
-
-        <div class="page-content">
-            <div class="contact-container">
-                <div class="contact-info">
-                    <h2>
-                        <span class="lang-es">Información de contacto</span>
-                        <span class="lang-ca">Informació de contacte</span>
-                        <span class="lang-en">Contact information</span>
-                    </h2>
-                    <p><i class="fas fa-map-marker-alt"></i> Aribau 271 bajos, 08021 Barcelona (España)</p>
-                    <p><i class="fas fa-phone"></i> 931 249 086</p>
-
-                    <h3>
-                        <span class="lang-es">Horario</span>
-                        <span class="lang-ca">Horari</span>
-                        <span class="lang-en">Hours</span>
-                    </h3>
-                    <p>
-                        <span class="lang-es">Martes a Viernes: 10:00 - 19:00</span>
-                        <span class="lang-ca">Dimarts a Divendres: 10:00 - 19:00</span>
-                        <span class="lang-en">Tuesday to Friday: 10:00 - 19:00</span>
-                    </p><p>
-                        <span class="lang-es">Sábado, Domingo y Lunes: Cerrado</span>
-                        <span class="lang-ca">Dissabte, Diumenge i Dilluns: Tancat</span>
-                        <span class="lang-en">Saturday, Sunday and Monday: Closed</span>
-                    </p>
-                    <div class="map-container">
-                        <h3>
-                            <span class="lang-es">Nuestra ubicación</span>
-                            <span class="lang-ca">La nostra ubicació</span>
-                            <span class="lang-en">Our location</span>
-                        </h3>
-                        <div class="transport-info">
-                            <p><i class="fas fa-bus"></i> <span class="lang-es">Línea V13 - Parada más cercana: Aribau - Laforja (a 4 min a pie)</span>
-                               <span class="lang-ca">Línia V13 - Parada més propera: Aribau - Laforja (a 4 min a peu)</span>
-                               <span class="lang-en">Line V13 - Nearest stop: Aribau - Laforja (4 min walk)</span>
-                            </p>
-                            <p><i class="fas fa-train"></i> <span class="lang-es">FGC: Estaciones Muntaner, Sant Gervasi y Plaça Molina (a unos 10-13 min a pie)</span>
-                               <span class="lang-ca">FGC: Estacions Muntaner, Sant Gervasi i Plaça Molina (a uns 10-13 min a peu)</span>
-                               <span class="lang-en">FGC: Muntaner, Sant Gervasi and Plaça Molina stations (about 10-13 min walk)</span>
-                            </p>
-                        </div>
-                        <iframe src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2993.874982!2d2.144664!3d41.394897!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x12a4a2f3f!2sAribau%20271%20bajos%2C%2008021%20Barcelona%2C%20España!5e0!3m2!1ses!2ses!4v1633036800000!5m2!1ses!2ses" width="600" height="300" style="border:0;" allowfullscreen="" loading="lazy"></iframe>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </main>
-
-    <!-- Pie de página -->
-    <footer>
-        <div class="footer-content">
-            <div class="footer-info">
-                <h3>BellaDona</h3>
-                <p><i class="fas fa-map-marker-alt"></i> Aribau 271 bajos, 08021 Barcelona</p>
-                <p><i class="fas fa-phone"></i> 931 249 086</p>
-            </div>
-            <div class="footer-social">
-                <a href="#" aria-label="Facebook"><i class="fab fa-facebook-f"></i></a>
-                <a href="#" aria-label="Instagram"><i class="fab fa-instagram"></i></a>
-                <a href="#" aria-label="Twitter"><i class="fab fa-twitter"></i></a>
-            </div>
-        </div>
-        <div class="footer-bottom">
-            <p>&copy; 2023 BellaDona. <span class="lang-es">Todos los derechos reservados.</span><span class="lang-ca">Tots els drets reservats.</span><span class="lang-en">All rights reserved.</span></p>
-        </div>
-    </footer>
-
-    <!-- Enlace a nuestro archivo JavaScript -->
-    <script src="script.js"></script>
-</body>
-</html>
