--- css/faq-contact.css
@@ -1,330 +1,0 @@
-/* FAQ Section */
-.faq {
-    background-color: white;
-    position: relative;
-    overflow: hidden;
-}
-
-.faq .section-header h2 {
-    color: var(--primary-color);
-}
-
-.faq-container {
-    max-width: 800px;
-    margin: 0 auto;
-}
-
-.faq-item {
-    margin-bottom: 15px;
-    border-radius: var(--border-radius);
-    overflow: hidden;
-    box-shadow: var(--box-shadow);
-    transition: all 0.3s ease;
-    background-color: white;
-    border: 1px solid #eee;
-}
-
-.faq-item:hover {
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
-}
-
-.faq-question {
-    width: 100%;
-    padding: 20px 25px;
-    text-align: left;
-    background: none;
-    border: none;
-    font-weight: 600;
-    font-size: 1.1rem;
-    color: var(--text-color);
-    cursor: pointer;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    transition: all 0.3s ease;
-    margin: 0;
-}
-
-.faq-question:hover {
-    color: var(--primary-color);
-}
-
-.faq-icon {
-    width: 24px;
-    height: 24px;
-    position: relative;
-    flex-shrink: 0;
-    margin-left: 15px;
-}
-
-.faq-icon::before,
-.faq-icon::after {
-    content: '';
-    position: absolute;
-    background-color: var(--primary-color);
-    transition: all 0.3s ease;
-}
-
-.faq-icon::before {
-    width: 16px;
-    height: 2px;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.faq-icon::after {
-    width: 2px;
-    height: 16px;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.faq-item.active .faq-icon::after {
-    opacity: 0;
-    transform: translate(-50%, -50%) rotate(90deg);
-}
-
-.faq-answer {
-    max-height: 0;
-    overflow: hidden;
-    transition: max-height 0.3s ease-in-out;
-    color: #666;
-    line-height: 1.7;
-    border-top: 1px solid transparent;
-    padding: 0 25px;
-}
-
-.faq-answer-inner {
-    padding: 0 0 25px;
-    color: #555;
-    line-height: 1.6;
-}
-
-.faq-item.active .faq-answer {
-    border-top-color: #eee;
-    padding: 0 25px;
-}
-
-/* Hover states */
-.faq-item:hover {
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
-}
-
-.faq-question:hover {
-    color: var(--primary-color);
-}
-
-/* Focus styles for accessibility */
-.faq-question:focus {
-    outline: 2px solid var(--primary-color);
-    outline-offset: 2px;
-}
-
-/* Contact Section */
-.contact {
-    background-color: var(--light-bg);
-    position: relative;
-    overflow: hidden;
-}
-
-.contact .section-header h2 {
-    color: var(--primary-color);
-}
-
-.contact-container {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 40px;
-    margin-top: 40px;
-}
-
-.contact-info {
-    background: white;
-    padding: 40px;
-    border-radius: var(--border-radius);
-    box-shadow: var(--box-shadow);
-}
-
-.contact-info h3 {
-    color: var(--primary-color);
-    margin-bottom: 25px;
-    position: relative;
-    padding-bottom: 15px;
-}
-
-.contact-info h3::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 50px;
-    height: 3px;
-    background-color: var(--accent-color);
-}
-
-.contact-details {
-    margin-bottom: 30px;
-}
-
-.contact-details p {
-    margin-bottom: 15px;
-    display: flex;
-    align-items: flex-start;
-    gap: 15px;
-}
-
-.contact-details i {
-    color: var(--accent-color);
-    font-size: 1.2rem;
-    margin-top: 3px;
-    min-width: 20px;
-    text-align: center;
-}
-
-.contact-details a {
-    color: var(--text-color);
-    transition: var(--transition);
-}
-
-.contact-details a:hover {
-    color: var(--primary-color);
-    text-decoration: none;
-}
-
-.opening-hours h4 {
-    color: var(--primary-color);
-    margin: 30px 0 15px;
-    position: relative;
-    padding-bottom: 10px;
-}
-
-.opening-hours h4::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 35px;
-    height: 2px;
-    background-color: var(--accent-color);
-}
-
-.opening-hours p {
-    margin-bottom: 8px;
-    display: flex;
-    justify-content: space-between;
-    max-width: 250px;
-}
-
-.opening-hours span:last-child {
-    font-weight: 600;
-    color: var(--primary-color);
-}
-
-.contact-form {
-    background: white;
-    padding: 40px;
-    border-radius: var(--border-radius);
-    box-shadow: var(--box-shadow);
-}
-
-.form-group {
-    margin-bottom: 20px;
-}
-
-.form-group label {
-    display: block;
-    margin-bottom: 8px;
-    font-weight: 500;
-    color: var(--text-color);
-}
-
-.form-control {
-    width: 100%;
-    padding: 12px 15px;
-    border: 1px solid #ddd;
-    border-radius: var(--border-radius);
-    font-family: inherit;
-    font-size: 1rem;
-    transition: var(--transition);
-}
-
-textarea.form-control {
-    min-height: 150px;
-    resize: vertical;
-}
-
-.form-control:focus {
-    outline: none;
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 2px rgba(194, 24, 91, 0.2);
-}
-
-.submit-btn {
-    background-color: var(--primary-color);
-    color: white;
-    border: none;
-    padding: 12px 30px;
-    border-radius: 50px;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    cursor: pointer;
-    transition: var(--transition);
-    width: 100%;
-}
-
-.submit-btn:hover {
-    background-color: var(--accent-color);
-    transform: translateY(-2px);
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
-}
-
-/* Map */
-.map-container {
-    height: 400px;
-    border-radius: var(--border-radius);
-    overflow: hidden;
-    box-shadow: var(--box-shadow);
-    margin-top: 40px;
-}
-
-.map-container iframe {
-    width: 100%;
-    height: 100%;
-    border: none;
-}
-
-/* Responsive Styles */
-@media (max-width: 992px) {
-    .contact-container {
-        grid-template-columns: 1fr;
-        max-width: 700px;
-        margin: 0 auto;
-    }
-    
-    .map-container {
-        margin-top: 20px;
-    }
-}
-
-@media (max-width: 768px) {
-    .contact-info,
-    .contact-form {
-        padding: 30px 20px;
-    }
-    
-    .faq-question {
-        padding: 15px 20px;
-        font-size: 1rem;
-    }
-    
-    .faq-answer {
-        padding: 0 20px;
-    }
-    
-    .faq-answer-inner {
-        padding: 0 0 20px;
-    }
-}
