--- css/footer-utilities.css
@@ -1,395 +1,0 @@
-/* Footer */
-.footer {
-    background-color: var(--dark-bg);
-    color: #bbb;
-    padding: 80px 0 0;
-    position: relative;
-    overflow: hidden;
-}
-
-.footer::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 5px;
-    background: linear-gradient(90deg, var(--primary-color), var(--accent-color));
-}
-
-.footer-top {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 40px;
-    margin-bottom: 60px;
-}
-
-.footer-about h3 {
-    color: white;
-    font-size: 1.8rem;
-    margin-bottom: 20px;
-    position: relative;
-    display: inline-block;
-}
-
-.footer-about h3::after {
-    content: '';
-    position: absolute;
-    bottom: -10px;
-    left: 0;
-    width: 50px;
-    height: 3px;
-    background-color: var(--accent-color);
-}
-
-.footer-about p {
-    margin-bottom: 20px;
-    line-height: 1.7;
-}
-
-.footer-links h4,
-.footer-newsletter h4 {
-    color: white;
-    font-size: 1.3rem;
-    margin-bottom: 25px;
-    position: relative;
-    padding-bottom: 10px;
-}
-
-.footer-links h4::after,
-.footer-newsletter h4::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 40px;
-    height: 2px;
-    background-color: var(--accent-color);
-}
-
-.footer-links ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.footer-links li {
-    margin-bottom: 12px;
-}
-
-.footer-links a {
-    color: #bbb;
-    transition: var(--transition);
-    display: inline-block;
-    position: relative;
-    padding-left: 0;
-}
-
-.footer-links a::before {
-    content: '→';
-    position: absolute;
-    left: -20px;
-    opacity: 0;
-    transition: var(--transition);
-    color: var(--accent-color);
-}
-
-.footer-links a:hover {
-    color: white;
-    padding-left: 20px;
-    text-decoration: none;
-}
-
-.footer-links a:hover::before {
-    left: 0;
-    opacity: 1;
-}
-
-.social-links {
-    display: flex;
-    gap: 15px;
-    margin-top: 20px;
-}
-
-.social-links a {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 40px;
-    height: 40px;
-    background-color: rgba(255, 255, 255, 0.05);
-    color: white;
-    border-radius: 50%;
-    transition: var(--transition);
-}
-
-.social-links a:hover {
-    background-color: var(--accent-color);
-    transform: translateY(-3px);
-    text-decoration: none;
-}
-
-.newsletter-form {
-    display: flex;
-    margin-top: 20px;
-}
-
-.newsletter-input {
-    flex: 1;
-    padding: 12px 15px;
-    border: none;
-    border-radius: 30px 0 0 30px;
-    font-family: inherit;
-    font-size: 0.95rem;
-    background-color: rgba(255, 255, 255, 0.05);
-    color: white;
-    transition: var(--transition);
-}
-
-.newsletter-input:focus {
-    outline: none;
-    background-color: rgba(255, 255, 255, 0.1);
-}
-
-.newsletter-btn {
-    background-color: var(--accent-color);
-    color: white;
-    border: none;
-    padding: 0 20px;
-    border-radius: 0 30px 30px 0;
-    cursor: pointer;
-    transition: var(--transition);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.newsletter-btn:hover {
-    background-color: var(--primary-color);
-}
-
-.footer-bottom {
-    border-top: 1px solid rgba(255, 255, 255, 0.05);
-    padding: 25px 0;
-    text-align: center;
-}
-
-.copyright {
-    margin: 0;
-    font-size: 0.9rem;
-    color: #777;
-}
-
-.copyright a {
-    color: var(--accent-color);
-    text-decoration: none;
-    transition: var(--transition);
-}
-
-.copyright a:hover {
-    text-decoration: underline;
-}
-
-/* Back to Top Button */
-.back-to-top {
-    position: fixed;
-    bottom: 30px;
-    right: 30px;
-    width: 50px;
-    height: 50px;
-    background-color: var(--accent-color);
-    color: white;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 1.2rem;
-    cursor: pointer;
-    opacity: 0;
-    visibility: hidden;
-    transition: var(--transition);
-    z-index: 999;
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.2);
-}
-
-.back-to-top.visible {
-    opacity: 1;
-    visibility: visible;
-}
-
-.back-to-top:hover {
-    background-color: var(--primary-color);
-    transform: translateY(-3px);
-}
-
-/* Responsive Styles */
-@media (max-width: 992px) {
-    .footer-top {
-        grid-template-columns: 1fr 1fr;
-    }
-    
-    .footer-newsletter {
-        grid-column: span 2;
-        max-width: 500px;
-    }
-}
-
-@media (max-width: 768px) {
-    .footer {
-        padding: 60px 0 0;
-    }
-    
-    .footer-top {
-        grid-template-columns: 1fr;
-        gap: 30px;
-    }
-    
-    .footer-newsletter {
-        grid-column: auto;
-        max-width: 100%;
-    }
-    
-    .footer-about,
-    .footer-links,
-    .footer-newsletter {
-        text-align: center;
-    }
-    
-    .footer-about h3::after,
-    .footer-links h4::after,
-    .footer-newsletter h4::after {
-        left: 50%;
-        transform: translateX(-50%);
-    }
-    
-    .social-links {
-        justify-content: center;
-    }
-    
-    .back-to-top {
-        width: 45px;
-        height: 45px;
-        bottom: 20px;
-        right: 20px;
-    }
-}
-
-/* Utility Classes */
-.text-center {
-    text-align: center;
-}
-
-.mb-0 {
-    margin-bottom: 0 !important;
-}
-
-.mt-0 {
-    margin-top: 0 !important;
-}
-
-.py-0 {
-    padding-top: 0 !important;
-    padding-bottom: 0 !important;
-}
-
-.pb-0 {
-    padding-bottom: 0 !important;
-}
-
-.pt-0 {
-    padding-top: 0 !important;
-}
-
-.container-sm {
-    max-width: 800px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.section-sm {
-    padding: 60px 0;
-}
-
-.bg-light {
-    background-color: var(--light-bg) !important;
-}
-
-.bg-dark {
-    background-color: var(--dark-bg) !important;
-    color: white;
-}
-
-.bg-primary {
-    background-color: var(--primary-color) !important;
-    color: white;
-}
-
-.bg-accent {
-    background-color: var(--accent-color) !important;
-    color: white;
-}
-
-.text-primary {
-    color: var(--primary-color) !important;
-}
-
-.text-accent {
-    color: var(--accent-color) !important;
-}
-
-.btn-block {
-    display: block;
-    width: 100%;
-}
-
-/* Animations */
-@keyframes fadeIn {
-    from {
-        opacity: 0;
-    }
-    to {
-        opacity: 1;
-    }
-}
-
-@keyframes slideInUp {
-    from {
-        opacity: 0;
-        transform: translateY(30px);
-    }
-    to {
-        opacity: 1;
-        transform: translateY(0);
-    }
-}
-
-.animate-fade-in {
-    animation: fadeIn 0.6s ease-out forwards;
-}
-
-.animate-slide-up {
-    animation: slideInUp 0.6s ease-out forwards;
-}
-
-/* Print Styles */
-@media print {
-    .no-print {
-        display: none !important;
-    }
-    
-    body {
-        padding: 20px;
-        color: #000;
-        background: #fff;
-    }
-    
-    a[href^="http"]::after {
-        content: " (" attr(href) ")";
-    }
-    
-    .container {
-        width: 100%;
-        max-width: 100%;
-        padding: 0;
-    }
-}
