--- css/gallery-testimonials.css
@@ -1,215 +1,0 @@
-/* Gallery Section */
-.gallery {
-    background-color: var(--light-bg);
-    position: relative;
-    overflow: hidden;
-}
-
-.gallery .section-header h2 {
-    color: var(--primary-color);
-}
-
-.gallery-grid {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 15px;
-    margin-top: 40px;
-}
-
-.gallery-item {
-    position: relative;
-    border-radius: var(--border-radius);
-    overflow: hidden;
-    aspect-ratio: 1;
-    cursor: pointer;
-    transition: var(--transition);
-}
-
-.gallery-item::before {
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(to bottom, rgba(0,0,0,0.1), rgba(0,0,0,0.7));
-    opacity: 0;
-    transition: var(--transition);
-    z-index: 1;
-}
-
-.gallery-item img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    transition: var(--transition);
-}
-
-.gallery-item:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
-}
-
-.gallery-item:hover::before {
-    opacity: 1;
-}
-
-.gallery-item:hover img {
-    transform: scale(1.1);
-}
-
-.gallery-caption {
-    position: absolute;
-    bottom: -50px;
-    left: 0;
-    width: 100%;
-    padding: 15px;
-    background: rgba(0, 0, 0, 0.7);
-    color: white;
-    transition: var(--transition);
-    z-index: 2;
-}
-
-.gallery-item:hover .gallery-caption {
-    bottom: 0;
-}
-
-/* Testimonials Section */
-.testimonials {
-    background: linear-gradient(rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.8)), 
-                url('../img/testimonial-bg.jpg') no-repeat center center/cover;
-    color: white;
-    position: relative;
-    text-align: center;
-}
-
-.testimonials .section-header h2,
-.testimonials .section-header .subtitle {
-    color: white;
-}
-
-.testimonials .section-header h2::after {
-    background-color: var(--accent-color);
-}
-
-.testimonials-slider {
-    max-width: 800px;
-    margin: 0 auto;
-    position: relative;
-}
-
-.testimonial-card {
-    background: rgba(255, 255, 255, 0.1);
-    backdrop-filter: blur(10px);
-    border-radius: var(--border-radius);
-    padding: 40px;
-    margin: 20px;
-    box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2);
-    transition: var(--transition);
-}
-
-.testimonial-avatar {
-    width: 80px;
-    height: 80px;
-    border-radius: 50%;
-    margin: 0 auto 20px;
-    border: 3px solid var(--accent-color);
-    overflow: hidden;
-}
-
-.testimonial-avatar img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.testimonial-text {
-    font-size: 1.1rem;
-    line-height: 1.8;
-    margin-bottom: 20px;
-    font-style: italic;
-    position: relative;
-}
-
-.testimonial-text::before,
-.testimonial-text::after {
-    content: '"';
-    font-size: 3rem;
-    color: var(--accent-color);
-    opacity: 0.5;
-    font-family: Georgia, serif;
-    line-height: 1;
-    position: absolute;
-}
-
-.testimonial-text::before {
-    top: -20px;
-    left: -10px;
-}
-
-.testimonial-text::after {
-    bottom: -40px;
-    right: -10px;
-}
-
-.testimonial-author {
-    font-weight: 700;
-    color: var(--accent-color);
-    margin-bottom: 5px;
-    font-size: 1.2rem;
-}
-
-.testimonial-role {
-    color: #ccc;
-    font-size: 0.9rem;
-}
-
-/* Slider Navigation */
-.slider-nav {
-    display: flex;
-    justify-content: center;
-    margin-top: 30px;
-    gap: 10px;
-}
-
-.slider-dot {
-    width: 12px;
-    height: 12px;
-    border-radius: 50%;
-    background-color: rgba(255, 255, 255, 0.3);
-    cursor: pointer;
-    transition: var(--transition);
-}
-
-.slider-dot.active {
-    background-color: var(--accent-color);
-    transform: scale(1.2);
-}
-
-/* Responsive Styles */
-@media (max-width: 992px) {
-    .gallery-grid {
-        grid-template-columns: repeat(2, 1fr);
-    }
-    
-    .testimonial-card {
-        padding: 30px 20px;
-    }
-}
-
-@media (max-width: 768px) {
-    .gallery-grid {
-        grid-template-columns: 1fr;
-        max-width: 500px;
-        margin: 0 auto;
-    }
-    
-    .testimonial-text {
-        font-size: 1rem;
-    }
-    
-    .testimonial-avatar {
-        width: 70px;
-        height: 70px;
-    }
-}
