--- css/hero-custom.css
@@ -1,255 +1,0 @@
-/* Hero Section Custom Styles */
-.hero {
-    min-height: 100vh;
-    background: linear-gradient(135deg, rgba(0, 0, 0, 0.7), rgba(0, 0, 0, 0.5)), url('../img/hero-bg.jpg');
-    background-size: cover;
-    background-position: center;
-    background-attachment: fixed;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
-    color: white;
-    position: relative;
-}
-
-.hero-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, 0.4);
-    z-index: 1;
-}
-
-.hero-content {
-    position: relative;
-    z-index: 2;
-    max-width: 800px;
-    margin: 0 auto;
-    padding: 0 20px;
-}
-
-.hero-logo-container {
-    margin-bottom: 30px;
-}
-
-.hero-logo-icon {
-    filter: drop-shadow(0 4px 8px rgba(0, 0, 0, 0.3));
-}
-
-.hero-text h1 {
-    font-size: 3.5rem;
-    font-weight: 700;
-    margin-bottom: 30px;
-    text-shadow: 2px 2px 4px rgba(0, 0, 0, 0.5);
-}
-
-.hero-features {
-    margin-bottom: 40px;
-    text-align: left;
-    max-width: 600px;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-.feature-item {
-    display: flex;
-    align-items: center;
-    margin-bottom: 15px;
-    font-size: 1.1rem;
-}
-
-.feature-icon {
-    color: #d4af37;
-    font-weight: bold;
-    margin-right: 15px;
-    font-size: 1.2rem;
-}
-
-.hero-cta {
-    margin-top: 40px;
-}
-
-.cta-button.summer-promo {
-    background: linear-gradient(45deg, #ff6b35, #f7931e);
-    color: white;
-    padding: 18px 40px;
-    font-size: 1.2rem;
-    font-weight: 700;
-    text-decoration: none;
-    border-radius: 50px;
-    display: inline-block;
-    transition: all 0.3s ease;
-    box-shadow: 0 8px 25px rgba(255, 107, 53, 0.3);
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    animation: pulse 2s infinite;
-}
-
-.cta-button.summer-promo:hover {
-    transform: translateY(-3px);
-    box-shadow: 0 12px 35px rgba(255, 107, 53, 0.4);
-}
-
-@keyframes pulse {
-    0% {
-        box-shadow: 0 8px 25px rgba(255, 107, 53, 0.3);
-    }
-    50% {
-        box-shadow: 0 8px 25px rgba(255, 107, 53, 0.5);
-    }
-    100% {
-        box-shadow: 0 8px 25px rgba(255, 107, 53, 0.3);
-    }
-}
-
-/* Services Section Custom Styles */
-.services {
-    padding: 80px 0;
-    background: #f8f9fa;
-}
-
-.services-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-    gap: 40px;
-    margin-top: 60px;
-}
-
-.service-card {
-    background: white;
-    padding: 40px 30px;
-    border-radius: 15px;
-    text-align: center;
-    box-shadow: 0 10px 30px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s ease;
-    border: 1px solid #e9ecef;
-}
-
-.service-card:hover {
-    transform: translateY(-10px);
-    box-shadow: 0 20px 40px rgba(0, 0, 0, 0.15);
-}
-
-.service-icon {
-    color: #d4af37;
-    margin-bottom: 25px;
-}
-
-.service-content h3 {
-    font-size: 1.5rem;
-    font-weight: 600;
-    margin-bottom: 15px;
-    color: #333;
-}
-
-.service-content p {
-    color: #666;
-    line-height: 1.6;
-    margin-bottom: 0;
-}
-
-/* Footer Custom Styles */
-.footer {
-    background: #333;
-    color: white;
-    padding: 40px 0;
-    text-align: center;
-}
-
-.footer-content {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    max-width: 1200px;
-    margin: 0 auto;
-}
-
-.footer-info h3 {
-    font-size: 1.8rem;
-    margin-bottom: 15px;
-    color: #d4af37;
-}
-
-.footer-info p {
-    margin-bottom: 10px;
-    color: #ccc;
-}
-
-.footer-info a {
-    color: white;
-    text-decoration: none;
-}
-
-.footer-info a:hover {
-    color: #d4af37;
-}
-
-.footer-social {
-    display: flex;
-    align-items: center;
-}
-
-.social-links {
-    display: flex;
-    gap: 15px;
-}
-
-.social-link {
-    width: 40px;
-    height: 40px;
-    border-radius: 8px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    text-decoration: none;
-    color: white;
-    font-weight: bold;
-    transition: all 0.3s ease;
-}
-
-.social-link.facebook {
-    background: #3b5998;
-}
-
-.social-link.instagram {
-    background: linear-gradient(45deg, #f09433, #e6683c, #dc2743, #cc2366, #bc1888);
-}
-
-.social-link.google {
-    background: #dd4b39;
-}
-
-.social-link:hover {
-    transform: translateY(-3px);
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
-}
-
-/* Responsive Design */
-@media (max-width: 768px) {
-    .hero-text h1 {
-        font-size: 2.5rem;
-    }
-    
-    .feature-item {
-        font-size: 1rem;
-    }
-    
-    .cta-button.summer-promo {
-        font-size: 1rem;
-        padding: 15px 30px;
-    }
-    
-    .footer-content {
-        flex-direction: column;
-        gap: 30px;
-    }
-    
-    .services-grid {
-        grid-template-columns: 1fr;
-        gap: 30px;
-    }
-}
-
