--- css/styles.backup.css
@@ -1,226 +1,0 @@
-/* Main CSS Import File */
-@import url('base.css');
-@import url('hero-services.css');
-@import url('gallery-testimonials.css');
-@import url('faq-contact.css');
-@import url('footer-utilities.css');
-@import url('hero-custom.css');
-
-/* Navbar */
-.navbar {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    background-color: rgba(255, 255, 255, 0.95);
-    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-    z-index: 1000;
-    padding: 15px 0;
-    transition: var(--transition);
-}
-
-.navbar.scrolled {
-    padding: 10px 0;
-    background-color: rgba(255, 255, 255, 0.98);
-    box-shadow: 0 4px 15px rgba(0, 0, 0, 0.1);
-}
-
-.navbar .container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: relative;
-}
-
-.nav-container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    justify-content: space-between;
-}
-
-.main-navigation {
-    display: flex;
-    align-items: center;
-}
-
-.logo {
-    font-size: 1.8rem;
-    font-weight: 700;
-    color: var(--primary-color);
-    text-decoration: none;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.logo img {
-    height: 40px;
-    width: auto;
-}
-
-.menu {
-    display: flex;
-    list-style: none;
-    margin: 0;
-    padding: 0;
-    gap: 20px;
-    align-items: center;
-}
-
-.menu-link {
-    color: var(--text-color);
-    font-weight: 500;
-    position: relative;
-    font-size: 0.95rem;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    text-decoration: none;
-    padding: 10px 0;
-    transition: var(--transition);
-}
-
-.menu-link::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 0;
-    height: 2px;
-    background-color: var(--primary-color);
-    transition: var(--transition);
-}
-
-.menu-link:hover,
-.menu-link.active {
-    color: var(--primary-color);
-}
-
-.menu-link:hover::after,
-.menu-link.active::after {
-    width: 100%;
-}
-
-.language-selector {
-    display: flex;
-    gap: 8px;
-    margin-left: 30px;
-    align-items: center;
-}
-
-.lang-btn {
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 5px;
-    border-radius: 50%;
-    transition: var(--transition);
-    width: 32px;
-    height: 32px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 1px solid transparent;
-}
-
-.lang-btn:hover {
-    transform: scale(1.1);
-    border-color: var(--primary-color);
-}
-
-.lang-btn.active {
-    border-color: var(--primary-color);
-    box-shadow: 0 0 0 2px var(--accent-color);
-}
-
-.lang-btn img {
-    width: 100%;
-    height: 100%;
-    border-radius: 50%;
-    object-fit: cover;
-}
-
-.hamburger {
-    display: none;
-    flex-direction: column;
-    justify-content: space-between;
-    width: 30px;
-    height: 21px;
-    background: none;
-    border: none;
-    cursor: pointer;
-    padding: 0;
-    z-index: 1001;
-}
-
-.hamburger span {
-    display: block;
-    width: 100%;
-    height: 3px;
-    background-color: var(--primary-color);
-    transition: var(--transition);
-    border-radius: 3px;
-}
-
-.hamburger.active span:nth-child(1) {
-    transform: translateY(9px) rotate(45deg);
-}
-
-.hamburger.active span:nth-child(2) {
-    opacity: 0;
-}
-
-.hamburger.active span:nth-child(3) {
-    transform: translateY(-9px) rotate(-45deg);
-}
-
-/* Responsive Navbar */
-@media (max-width: 992px) {
-    .hamburger {
-        display: flex;
-    }
-    
-    .main-navigation {
-        position: fixed;
-        top: 0;
-        right: -100%;
-        width: 80%;
-        max-width: 320px;
-        height: 100vh;
-        background-color: white;
-        flex-direction: column;
-        justify-content: flex-start;
-        padding-top: 100px;
-        gap: 0;
-        box-shadow: -5px 0 15px rgba(0, 0, 0, 0.1);
-        padding: 100px 20px 40px;
-        z-index: 1000;
-        transition: right 0.3s ease-in-out;
-    }
-    
-    .main-navigation.active {
-        right: 0;
-    }
-    
-    .menu {
-        flex-direction: column;
-        width: 100%;
-        gap: 0;
-    }
-    
-    .menu li {
-        width: 100%;
-        text-align: center;
-    }
-    
-    .menu-link {
-        display: block;
-        padding: 15px 0;
-        font-size: 1.1rem;
-    }
-    
-    .language-selector {
-        margin: 30px auto 0;
-        justify-content: center;
-    }
-}
