+++ index.html
@@ -33,7 +33,7 @@
     <link rel="canonical" href="https://peluqueria-belladona.netlify.app/">
     <meta name="google-site-verification" content="XScpcggOWd--tlfwg-T14iPSZOJkWh-dnb-ttPrle0g" />
     
-    <title>BellaDona - Peluquería Profesional</title>
+    <title>BellaDona - Peluquería Profesional en Barcelona</title>
     <link rel="stylesheet" href="css/styles.css">
     <link rel="stylesheet" href="css/faq-contact.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
@@ -225,7 +225,7 @@
         <div class="container">
             <div class="hero-content">
                 <div class="hero-text">
-                    <h1 data-i18n="hero.title">Tu belleza, nuestra pasión</h1>
+                    <h1 data-i18n="hero.title">BellaDona: Tu Peluquería en Sarrià-Sant Gervasi, Barcelona</h1>
                     <p class="hero-subtitle" data-i18n="hero.subtitle">Descubre la mejor experiencia de peluquería</p>
                     <a href="#reservas" class="btn btn-primary" data-i18n="hero.cta">Reserva ahora</a>
                     <div class="hero-features">
@@ -259,7 +259,7 @@
             </div>
         </div>
     </header>
-
+    <main id="main-content">
     <!-- Reservas Section -->
     <section id="reservas" class="reservas" aria-labelledby="reservas-heading">
         <div class="container">
@@ -491,7 +491,7 @@
             </div>
         </div>
     </section>
-
+    </main>
     <!-- Footer -->
     <footer class="footer">
         <div class="container">
@@ -542,8 +542,6 @@
     </script>
 
     <!-- Scripts en orden correcto: primero i18n, luego seasonal_cta, luego main -->
-    <script src="js/i18n.js"></script>
-    <script src="js/seasonal_cta.js"></script>
-    <script src="js/main.js"></script>
+    <script src="js/app.js"></script>
 </body>
 </html>
