--- netlify.toml
@@ -1,24 +1,0 @@
-[build]
-  publish = "."
-
-[[redirects]]
-  from = "/*"
-  to = "/index.html"
-  status = 200
-
-[[headers]]
-  for = "/*"
-  [headers.values]
-    Content-Security-Policy = "default-src 'self' https: data: 'unsafe-inline' 'unsafe-eval'; frame-src https://*.elfsight.com https://www.google.com; script-src 'self' https: 'unsafe-inline' 'unsafe-eval'; style-src 'self' https: 'unsafe-inline';"
-    X-Frame-Options = "ALLOW-FROM https://*.elfsight.com"
-    Referrer-Policy = "no-referrer-when-downgrade"
-    X-Content-Type-Options = "nosniff"
-    X-XSS-Protection = "1; mode=block"
-    Permissions-Policy = "geolocation=(), microphone=(), camera=()"
-
-
-[[redirects]]
-  from = "/*"
-  to = "/index.html"
-  status = 200
-
