--- robots.txt
@@ -1,5 +1,0 @@
-User-agent: *
-Allow: /
-
-Sitemap: https://peluqueria-belladona.netlify.app/sitemap.xml
-
