--- schema_code.html
@@ -1,72 +1,0 @@
-<script type="application/ld+json">
-{
-  "@context": "https://schema.org",
-  "@type": "HairSalon",
-  "name": "BellaDona Peluquería",
-  "image": "https://peluqueria-belladona.netlify.app/img/logo.png",
-  "@id": "",
-  "url": "https://peluqueria-belladona.netlify.app/",
-  "telephone": "+34931249086",
-  "priceRange": "€€",
-  "address": {
-    "@type": "PostalAddress",
-    "streetAddress": "Calle Aribau, 271",
-    "addressLocality": "Barcelona",
-    "postalCode": "08021",
-    "addressCountry": "ES"
-  },
-  "geo": {
-    "@type": "GeoCoordinates",
-    "latitude": 41.39861,
-    "longitude": 2.14861
-  },
-  "openingHoursSpecification": [
-    {
-      "@type": "OpeningHoursSpecification",
-      "dayOfWeek": [
-        "Tuesday",
-        "Wednesday",
-        "Thursday",
-        "Friday"
-      ],
-      "opens": "10:00",
-      "closes": "19:00"
-    }
-  ],
-  "sameAs": [
-    "https://www.instagram.com/belladona_peluqueria/"
-  ],
-  "description": "Peluquería profesional en Barcelona especializada en corte, color y tratamientos capilares. Con más de 15 años de experiencia, ofrecemos servicios personalizados en el barrio de Sarrià-Sant Gervasi.",
-  "hasOfferCatalog": {
-    "@type": "OfferCatalog",
-    "name": "Servicios de Peluquería",
-    "itemListElement": [
-      {
-        "@type": "Offer",
-        "itemOffered": {
-          "@type": "Service",
-          "name": "Corte y Peinado",
-          "description": "Cortes personalizados que realzan tu belleza natural"
-        }
-      },
-      {
-        "@type": "Offer",
-        "itemOffered": {
-          "@type": "Service",
-          "name": "Coloración",
-          "description": "Técnicas de color innovadoras para un look único"
-        }
-      },
-      {
-        "@type": "Offer",
-        "itemOffered": {
-          "@type": "Service",
-          "name": "Tratamientos",
-          "description": "Revitaliza tu cabello con nuestros tratamientos especializados"
-        }
-      }
-    ]
-  }
-}
-</script>
-
